samples/copilot-studio-extensibility/python/requirements.txt
durabletask-azuremanaged
azure-identity
azure-mgmt-cognitiveservices
requests
aiohttp
dataclasses-json
python-dotenv